Fix crash when trying to delete non-directories

+++ factor/scripts/delete_image.py
@@ -30,11 +30,13 @@ def main(imageroot, counter):
         images1  = glob.glob(imageroot + '.*')
         for image in images1:
             if os.path.exists(image):
-                shutil.rmtree(image)
+                if os.path.isdir(image)
+                    shutil.rmtree(image)
         images2  = glob.glob(imageroot.replace('image41', 'image42') + '.*')
         for image in images2:
             if os.path.exists(image):
-                shutil.rmtree(image)
+                if os.path.isdir(image)
+                    shutil.rmtree(image)
 
 
 if __name__ == '__main__':
